Move UTs as we target now net framework.
