--- Tests/SonarScanner.MSBuild.Tasks.NetFramework.UnitTest/Properties/AssemblyInfo.cs
@@ -1,27 +1,0 @@
-﻿/*
- * SonarScanner for .NET
- * Copyright (C) 2016-2021 SonarSource SA
- * mailto: info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("SonarScanner.MSBuild.Tasks.NetFramework.UnitTest")]
-[assembly: AssemblyProduct("SonarScanner.MSBuild.Tasks.NetFramework.UnitTest")]
-[assembly: AssemblyDescription("")]
-[assembly: Guid("997457c7-05f7-4cf9-a2ad-4c9c4460fe39")]
